Made use of the PostgreSQL plugin an explicit requirement
Changes to the structure of the Sonador plugin (quite a while ago) made
use of an external PostgreSQL database and the PostgreSQL plugin
necessary. This update disables the options that (sort of) allowed
people to try to by-pass it.

(Attempting to by-pass was a fool's errand, since the plugin
required the database.)

This commit just makes explicit what was implicit previously.